[Bugfix] DiffusionGemma: cast the self-conditioning soft embed to the buffer dtype
The soft embed is computed in the embedding weight's dtype and written
into the fp32 self-conditioning buffer. Inductor inserts the cast, so the
compiled step is fine, but once dynamo falls back to eager (past the
recompile limit, for instance when requests come in several canvas widths)
the index put raises a dtype mismatch and the engine dies. Cast explicitly
so both paths agree.

+++ vllm/model_executor/models/diffusion_gemma.py
@@ -649,7 +649,7 @@ def _compiled_sample_step(
                 soft_embeds, group_name=tp_group_name
             )
         soft_embeds = soft_embeds * normalizer
-        sc_embeds[decode_slots] = soft_embeds * sc_keep
+        sc_embeds[decode_slots] = (soft_embeds * sc_keep).to(sc_embeds.dtype)
     else:
         # Every slot in this tile ends after this step, so the soft embed
         # would never be read. The matmul is a full pass over the vocabulary
